chore(JSON-RPC): rename state mod to pending
