+++ crates/papyrus_rpc/src/lib.rs
@@ -5,10 +5,10 @@
 mod api;
 mod compression_utils;
 mod middleware;
+mod pending;
 mod rpc_metrics;
 #[cfg(test)]
 mod rpc_test;
-mod state;
 mod syncing_state;
 #[cfg(test)]
 mod test_utils;
